feat(reco): densify place detail page (UX3)
Render the place description ("Le lieu" block), add a photo gallery
mosaic, and make a single reco prominent so the detail page no longer
feels empty with one recommendation. Markup extracted into a pure
presentational component (RecommandationDetail) fed by the existing
SSR (anon/session) query path — no data/auth change.
